app/dependencies/auth.py
from fastapi import Depends, HTTPException, status
from sqlalchemy.orm import Session
from jwt import ExpiredSignatureError, InvalidTokenError
from app.crud import user as crud
from app.dependencies.db import get_db
from fastapi.security import OAuth2PasswordBearer

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

def get_current_user_safe(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
    try:
        current_user = crud.get_current_user(token, db)
    except ExpiredSignatureError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Token expired",
            headers={"WWW-Authenticate": "Bearer"},
        )
    except InvalidTokenError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid token",
            headers={"WWW-Authenticate": "Bearer"},
        )
    if current_user is None:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Not authenticated",
            headers={"WWW-Authenticate": "Bearer"},
        )
    return current_user