requirements.txt
alembic==1.16.2
bcrypt==4.3.0
email_validator==2.2.0
fastapi==0.115.14
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.7
PyJWT==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.41